@@ -1,6 +1,6 @@
 module github.com/opcr-io/policy
 
-go 1.16
+go 1.17
 
 // replace github.com/aserto-dev/clui => ../clui
 
